Changed the folder structure
