+++ src/components/emergeComponents/BulletPoint.jsx
@@ -1,5 +1,5 @@
 export default function BulletPoint(){
     return (
-        <div className="w-2 h-2 bg-blue-500 rounded-full mr-4 mt-2 flex-shrink-0"></div>
+        <div className="w-3.5 h-3.5 bg-fuchsia-500 rounded-full mr-5 mt-3 flex-shrink-0"></div>
     )
 }
