--- src/components/emergeComponents/Header.jsx
@@ -1,8 +1,0 @@
-export default function Header(){
-    <div className="text-center border-b border-zinc-800 pb-6 mb-8">
-        <h1 className="text-3xl sm:text-4xl font-extrabold text-white">E-MERGE '25 | Student Registration</h1>
-        <p className="mt-2 text-md text-gray-400">
-            IIT Hyderabad | 12th October 2025 | Organized by E-Cell, IIT Hyderabad
-        </p>
-    </div>
-}
