+++ src/components/emergeComponents/SectionHeader.jsx
@@ -1,7 +1,7 @@
 export default function SectionHeader({title}){
     return (
-        <div className="mb-6 mt-10 border-b border-zinc-700 pb-2">
-            <h2 className="text-2xl font-bold text-white">{title}</h2>
-        </div>
+        <div className="mb-12 mt-20 border-b border-gray-700 pb-6">
+        <h2 className="text-5xl font-bold text-white">{title}</h2>
+    </div>
     );
 };
